Create doorVisit.java
implements the behavior class to deliver pizza to specified house. Assumes that it starts on a road and will end at the head of a road.
Need info on how behavior class works. So far it seems that behaviors get activated/deactivated by certain conditions. This might be hard to code. Is this interpretation correct?